feat: move table logic for columns directly inside the page
