--- components/admin/users-datatable.tsx
@@ -1,119 +1,0 @@
-'use client'
-
-import type { ColumnDef } from '@tanstack/react-table'
-import { ArrowUpDown, MoreHorizontal } from 'lucide-react'
-
-import { Button } from '@/components/ui/button'
-import { Checkbox } from '@/components/ui/checkbox'
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger
-} from '@/components/ui/dropdown-menu'
-import { DataTable } from '@/components/ui/datatable'
-import { Loader } from '@/components/ui/shared/loader'
-import { Card, CardContent } from '@/components/ui/card'
-
-import { useUsers } from '@/hooks/mod/users/use-users'
-
-import { UserInformationPublicType } from '@/lib/definitions'
-
-export const columns: ColumnDef<UserInformationPublicType>[] = [
-  {
-    id: 'select',
-    header: ({ table }) => (
-      <Checkbox
-        checked={
-          table.getIsAllPageRowsSelected() || (table.getIsSomePageRowsSelected() && 'indeterminate')
-        }
-        onCheckedChange={(value) => table.toggleAllPageRowsSelected(!!value)}
-        aria-label='Select all'
-      />
-    ),
-    cell: ({ row }) => (
-      <Checkbox
-        checked={row.getIsSelected()}
-        onCheckedChange={(value) => row.toggleSelected(!!value)}
-        aria-label='Select row'
-      />
-    ),
-    enableSorting: false,
-    enableHiding: false
-  },
-  {
-    accessorKey: 'name',
-    header: 'Name',
-    cell: ({ row }) => <div className='capitalize'>{row.getValue('name')}</div>
-  },
-  {
-    accessorKey: 'email',
-    header: ({ column }) => {
-      return (
-        <Button
-          className='pl-0'
-          variant='ghost'
-          onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-        >
-          Email
-          <ArrowUpDown className='ml-2 h-4 w-4' />
-        </Button>
-      )
-    },
-    cell: ({ row }) => <div className='lowercase'>{row.getValue('email')}</div>
-  },
-  {
-    id: 'actions',
-    enableHiding: false,
-    cell: ({ row }) => {
-      const user = row.original
-
-      return (
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <Button variant='ghost' className='h-8 w-8 p-0'>
-              <span className='sr-only'>Open menu</span>
-              <MoreHorizontal className='h-4 w-4' />
-            </Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent align='end'>
-            <DropdownMenuLabel>Actions</DropdownMenuLabel>
-            <DropdownMenuSeparator />
-            <DropdownMenuItem>View customer</DropdownMenuItem>
-            <DropdownMenuItem>View payment details</DropdownMenuItem>
-          </DropdownMenuContent>
-        </DropdownMenu>
-      )
-    }
-  }
-]
-
-const UsersDatatable = () => {
-  const { loading, users } = useUsers()
-
-  return (
-    <>
-      {loading ? (
-        <div className='w-full h-[60vh] flex justify-center items-center'>
-          <Loader />
-        </div>
-      ) : (
-        <>
-          {Array.isArray(users) && users.length > 0 ? (
-            <DataTable data={users} columns={columns} />
-          ) : (
-            <Card className='p-8'>
-              <CardContent className='flex items-center justify-center pt-6'>
-                <div className='text-2xl font-bold'>Oops, no users has been joined for now</div>
-              </CardContent>
-            </Card>
-          )}
-        </>
-      )}
-    </>
-  )
-}
-
-export { UsersDatatable }
